Extract network-related stuff from AppModule
